Fixed an issue where I somehow managed to place the eurom tests in the geco folder and vice versa
